requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.0.0
django-cors-headers>=4.0.0
# PostgreSQL dependencies
python-decouple>=3.8
psycopg2-binary>=2.9.0
